Adds the idea of "local modules" to the source layout and build system.
A local module is one whose source is part of this project (as opposed
to coming in via npm) that is made available to both the client and
the server. The idea is to use these for any bits of code that are
_either_ actually used on both the client and the server, _or_ for
functionality that could reasonably be conceived of as being an
independent module (whether or not it happens to be used by both
client and server).

The three initial local modules are `delta-util`, `prom-condition`,
and `prom-delay`. The first is in fact used on both client and
server. The other two are not (but might indeed come in handy as such,
eventually).