Implement transform / rebase on both client and server.
On the server side, this now handles a request to apply a delta to something
other than the latest version, meaning that the client might now get a
non-empty "correction" delta in response to a submitted change.

On the client side, in addition to the easy cases of "correction" (which were
already handled), this now also allows the client to merge corrections
"underneath" ongoing changes. That is, the user can keep typing while a
change is in mid-flight, and the response coming back from the server will no
longer confuse the client.

Bumped version to 0.1.0, because this seems like a pretty legit milestone.