Adds a bottleneck local module `see-all` to handle logging.
Right now it's just a straightforward wrapper around `console.log()`,
but eventually it's where we can start to put some more smarts. This
is a prerequisite to cleaning up the spew currently being logged. That
is, I don't want to remove the logging calls, but I _do_ want to make
it easy to squelch them.

To be clear: I am not opposed to dropping in a more feature-ful
pre-existing logging library, but I don't want to do that _yet_.