Handle some errors.
The main point here is to get `DocumentPlumbing` to start to behave
somewhat sensibly in the face of a network trouble. That meant getting
`ApiClient` to provide more meaningful errors, such that the plumbing
can determine a bit better what's going on. There is still _way_ more
to do before we can count on the system to truly recover from network
interruptions, to be clear.

Bonus: Better browser logging stuff.