Handle local modules in dev mode.
When I introduced local modules, dev mode got left behind in that it
didn't understand how to synch those files. This change fixes that oversight.

What's going on here is that we now keep a file at the root of each directory
copied from the source tree into `out` which lists the source directories
from whence it came (including a primary directory and possibly an overlay).
The dev mode code finds all these files and uses them to figure out what
directories to watch and how to react when watched files changed.