Use Chokidar instead of Watchpack.
The main reason is that the former gives you a way to monitor file
removals but the latter does not.

As it turns out, Watchpack uses Chokidar under the covers, so this was
pretty straightforward. I _did_ have to rewrite Watchpack's convenient
"start looking for changes as of time N" where N can be a time in the
past; but that turned out to be reasonably easy.