Clean up the box building stuff.
This includes pointing us at `npmbox` v4.2.0, which I produced myself
with the blessing (and oversight) of the original author of the
package. There is a bit more we can do to take advantage of that
version, as noted in the comments; but for now I'm happy enough to
leave this as-is.