Verify Node version as a preflight before running.
This is a bit of code duplication from the similar build-time check, but
we don't currently have a place to put "script library" code which is common
between the build and run environments. So, duplication it shall be for now.
If we end up needing a lot more of the same code, perhaps it will be worth
factoring out, of course.