A bit more WIP on three-way introduction.
In this case, we now have a split-out JS file `boot-from-key.js` to
include (via `<script>`) in other HTML pages, which suffices to
bootstrap the editor front-end assuming you got an access key
(presumably via back-end communication). Another new file
`boot-from-key.html` is a stubbed-out use of this, which is suitable
for testing for the time being but isn't really the ultimate solution.

Still to come:

* Replace the main `index.html` with something that uses
  `boot-from-key.js`. It will probably be based on
  `boot-from-key.html`.

* Remove `boot-from-key.html`.

* Honor the document/author indicated by the access key. In the mean
  time, we just bottom out at using the same just-one-doc that we
  always have.