Top-level client unit test file, dynamically constructed.
Here's what's going on: We add a new file under `client/js/tests` called
simply `client-tests`, which isn't actually a JS file. Instead, its name is
listed explicitly in our Webpack config to trigger the use of a custom loader
called `client-tests-loader`. This loader is responsible for dynamically
constructing JS source that refers to all of our client-side test files.

As of this commit, the mechanism is in place for doing the synthesizing and
loading, but the generated file does not yet refer to any of the actual
test files. (See TODOs for a bit of an explanation why.)