Fix the "expected a document" bug.
This turned out to be because I wasn't providing the right delta to apply to
the client's version of the document when performing merges.

This bug had been around since pretty much day-1 of the synch code, but
(my best guess) it probably only started showing up when a Quill change
caused it to become more picky about document contents.