Make `changeAppend()` accept a `DocumentChange`.
It had previously just taken args to pass to the `DocumentChange` constructor,
because at the time it was possible to know what version number to construct
with, so that argument could be elided. However, in the mean time, the move
to an asynchronous interface meant that the version number did actually need
to be passed explicitly. With that, the implicit(ish) constructor call became
more obfuscatory than helpful.