Add retry logic to `applyDelta()`.
This handles the case where a call to the method is valid, but during the
course of its processing it ends up with an out-of-date snapshot. This can
happen when two or more users are actively editing a document; inevitably at
some point there are two "simultaneous enough" attepts to apply a delta, such
that their execution ends up overlapping. One of the attempts "wins the race"
and causes the document to be updated. Then, later, the other attempt "loses
the race" and discovers that the snapshot they got in the middle of handling
the call is no longer the right one to use.