Update the required browser versions, for ES6 `class` syntax support.
Babel cannot properly translate ES6 `class` syntax if the class in question
`extends` a built-in JS class, such as notably `Array` or `Error`, because
reasons. In fact, we define such a class in our code, `InfoError extends
Error`. This wasn't actually working, and was causing trouble.

As it turns out, recent versions of all the browsers we care about (and
Electron) now support `class` syntax, so it's just a matter of tweaking our
Babel rules, so that it no longer tries to translate `class`, instead just
passing it through as-is.