Fill in the caret methods.
This implementation should work as long as clients only ever ask to wait
for updates after the exact latest revision number. Ultimately, we will want it
to be able to handle requests with respect to the recent past (for some values
of the term "recent"), but this should be good enough for now (modulo further
debugging, natch).