Second attempt at caret storage into `file-store`.
This time, the storage part is handled by a separate class which operates
somewhat at arm's length to the more interactive `CaretControl`. The idea
here is that we can provide lower-latency interactivity in the common case
of editors operating via the same server, while still operating correctly
when editors span servers. In the latter case, we rate-limit how often carets
get pushed to storage.

This is still work-in-progress, because there still isn't any code to _read_
carets from storage.