Enforce a "median consensus" for a line length limit.
What this does is turn on a max length of 120 columns for code, but
limits comments to 80; and it waives the limits in all the common
cases where one is likely to end up wanting to violate the limit for
readability (e.g. literal strings).

Notably, with this rule imposed, there were only a small handful of
lines of code that were flagged by the linter.

Just to reiterate something I've said in person, the reason I want to
keep comment lines at 80 columns is to make it easier to read
paragraphs such as might be found in class and method header comments.