Less is more.
From @meantime:

This change does a bunch of things:

- adds the less-loader to our build system

- converts nearly all of our styling markup from css to less

- further takes our React component markup and puts it all into
  modularized files

- brings in new colors and styles

- replaces all of the ad hoc instances of those colors and styles in
  our markup and replaces them with the shared definitions

- one Less mixin helper is now there to manage the non-standard
  user-select tag.

There is less overall markup now, less redundancy in what is left, and
it's all more modular.

There are two rules for the less loader. One is for `.module.less`
files and one is for `.less` files. The former is being used with our
react components so that the component and its styling can travel
together. The JS code for the component actually imports the styling
file. The "module" part of it is that all of the class names in the
css are mangled during compilation to make them not able to interfere
with other classes. The non-"module" version is just straight less ->
css processing without the name mangling. This is used with the files
that have to interoperate with Quill and its classes (which we have
to, sadly).