First bits of redux document state... and an interaction!
From @meantime:

This adds a Redux data store to the client application. Currently
there are actions and reducers for managing the Owner and Star
sections of the page header. You can click the star now and it
toggles. Yay.

- ClientStore now present to compose the root reducer, and to wrap the
  constructed store.

- Reducers and mutation actions created for Star and Owner components.

- Star and Owner components updated to map redux state to its own props.

- Interaction added to the Star by means of an action mapping to
  mutate the redux state.

- Type validators added for Star and Owner properties.

I feel pretty good about the basic architecture of this whole
thing. Not sure on a few names, or on some of the file organization.