Initial work on embedding React components in Quill.
From @meantime:

- `ReactEmbed` is a new Quill `BlockEmbed` blot that can host React
  components.

- `EmbeddableComponent` is a base class for React components that you
  want to embed in Quill.

- There is a Clock component which is just a sample of a live-updating
  component that you can embed in a document.

`ReactEmbed` maintains a registry of embeddable components. The
components have to define a registry name, and a second set of
`PropTypes`. That second `PropTypes` declaration lists all of the
component's properties that you want to be saved/restored as part of
the document. Anything you put in there is automagically assigned as
data values on the DOM elements which Quill uses when flattening it
all out to a Delta, and are restored later when the Delta is inflated.