Redo of React embeds.
From @meantime:

* component registry is gone
* component base class is gone
* boilerplate steps to get a component embedded are gone

All the magic is now in the wrapper (which is dynamically generated)
instead of in a component base class. Now the embeds won't all share a
common name (which would have been a PITA for other clients to deal
with). h/t to @danfuzz for pointing out the problems with the previous
implementation.