Redux redux, from @meantime.
Update the style of all of our Redux code to conform with our local
corporate environment.

* All dispatch actions now have their data in a `payload` property.

* All dispatch actions created by functions.

* Selectors for all data access rather than digging into the state
  object like an animal.

* Unit tests for all of the modules that contribute to the redux
  client store.

The careful reader will note that the unit tests don't use the
selectors when teasing mutation results out of the state
objects. That's because the selectors were all written expecting to
receive a full state object (what you'd get back from
`clientStore.getState()`) but the tests are only testing the reducers
in isolation so they are receiving just the partial state object that
is managed by the given reducer.