First pass at implementing image embedding.
**Note:** There is currently no code available in the repo to test image
embeds, but if you manually inject one into a document, you can start to
play with it. We hope to rectify this deficiency soon.

From @meantime:

* Hovering over an image will allow you to choose left/center/right
  alignment and the rest of the document content should do its best to
  flow around it as you'd expect. There is a disabled option for a
  centered-full-bleed but I need help with the CSS for that.

* A handful of function hooks are passed down to embedded react
  components now so they can do things like push their props out to
  Quill, adjust the CSS classes on their container `DIV`, etc. These
  are needed because the embedded components are the root component
  and there's nothing above them to request that work from. There's
  lots of documentation in the code explaining what they all do and
  why they're needed.

The alignment toolbar is a total placeholder. It's meant to be a
popover toolbar above the image but it seemed better to wait and try
to find a way to import the webapp `popover` and `tooltip` components
in the future rather than make my own janky versions right now.

Anyway, here is a not-so-big PR that represents rewriting the world 3
times last week. There's some quill stuff, some react stuff, and some
css.