Reverse the test of what's found in `_complexes`.
As of this writing, there's a bug in either the `weak` module or perhaps in
the Node/V8 functionality that it relies on, whereby `weak.isWeakRef(x)` will
sometimes return `true` when `x` is not in fact a weak reference. In our case
it was returning `true` for an unresolved promise. So, instead of using that,
we check to see if what we have is a promise; if not, we assume it's a weak
reference.