Make the base `update()` do more of the work.
The remaining bit that gets implemented by the subclasses is `_impl_rebase()`,
which is _just_ the logic to rebase a change from its original base to a
given instantaneously-current base.

This commit fixes _some_ of the test cases and disables the two which are
rendered irrelevant (because of structural changes). A follow-up commit will
rework them.