Switch to using single-binding plain objects to represent instances.
This is instead of using "tagged" arrays.

Before:

* Instances encode as `['tag', <arg>, ...]`.
* Arrays encode as `['array', <elem>, ...]` (literal `array` first element).
* Plain objects self-encode.

Now:

* Instances encode as `{ tag: [<arg>, ...] }`.
* Arrays self-encode.
* Plain objects encode as `{ object: [[<k>, <v>], ...] }`.

The new instance encoding form was inspired by how Quill represents
the payloads for custom blots (it is pretty much recapitulated
exactly). The impetus was that our data encodes arrays way more than
it encodes plain objects, so it was a poor choice for arrays to be the
encoded form of instances.