Use `ConstructorCall`.
More specifically, we now differentiate between the "encoded data" form of
an object graph and the "JSON-encodable data" form of same. The former uses
instances of `ConstructorCall` to represent...well...constructor calls, and
the latter uses the one-binding plain object form established in an earlier
PR (and inspired by Quill's use of same).

The point of all this is to make it a bit easier to create a non-JSON
encoding form. In such a form it wouldn't be sensible to use that plain
object hack, and so it helps that it wouldn't end up having to either ingest
or produce such objects. The hack is now purely limited to JSON-specific
functionality.