Use a "default-safe idiom" for calling static methods.
Even though in these cases it's okay to call the method as a `this`-less
function, in many similar cases it's not. So, in order to (a) have
consistent-looking code, and (b) not lure the unsuspecting into doing the
unsafe thing, this commit switches references to `ClassName.methodName` passed
as arguments to instead be a closure `x => ClassName.methodName(x)`.