Fix module config calculation.
Not that this caused a real bug in practice, but it was bad form...
Specifically, prior to this change, variables that looked like constants
weren't actually being treated as such. Instead of trying to fix the call
sites to not treat the values in question as mutable, it was more
straightforward to just provide fresh mutable objects on each call.