DRY out the local module `package.json` files.
More specifically, remove `name` and `version`. Both are redundant, and the
latter is irrelevant for project-internal usage.

The long-term plan is to publish local modules to a scoped namespace, and
when that's done, the `package.json` files as published will be
script-generated instead of having to be checked in with lots of boilerplate.