Two fixes regarding validations.
From @vahakanp:

1. Fix the bug where migration and error messages would be appended to the end of any existing document.

2. Remove the last trace of `transact`.

The issue with 1. was that we were appending the migration/error note
with a revision number of `1`, so it would be buried behind all the
other revisions, thus not being applied to the document. This
manifested with the existing contents to not be deleted, but also with
the schema not being updated, etc.