Stop using `FileId.check()`.
`BaseFileStore` now ensures it constructs `BaseFile` instances with
valid IDs, so there's no need to do the extra check. And, more to the point,
with the ID validity API moving into `BaseFileStore` it's going to be more
trouble than it's worth to keep the redundant checks (even though I'm in favor
of this sort of "defense in depth" by default).