Stop removing targets from `Context` based on idle time.
This was essential in a world where there was a global registry of targets.
However, when each context is linked to a connection (or later, an API
session), the right thing to do is only remove targets (if ever) based on
what's happening at the connection (or API session) layer.