Make `EventSource.emit` a clever proxy.
In the pattern established by `Logger.event.*`, `EventSource.emit` is
a proxy which returns a function for (nearly) any requested property,
which when called emits the event named by the property. In addition,
it behaves as a direct function, implementing its old behavior of
taking a `Functor` which it emits as an event.